+++ core/src/main/java/org/sterl/spring/persistent_tasks/history/config/TriggerHistoryConfig.java
@@ -1,0 +1,20 @@
+package org.sterl.spring.persistent_tasks.history.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.TaskExecutor;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+@Configuration
+public class TriggerHistoryConfig {
+
+    @Bean
+    TaskExecutor triggerHistoryExecutor() {
+        var taskExecutor = new ThreadPoolTaskExecutor();
+        taskExecutor.setCorePoolSize(1);
+        taskExecutor.setMaxPoolSize(4);
+        taskExecutor.setWaitForTasksToCompleteOnShutdown(true);
+        taskExecutor.initialize();
+        return taskExecutor;
+    }
+}