+++ core/src/main/java/org/sterl/spring/persistent_tasks/trigger/component/RunTriggerComponent.java
@@ -1,6 +1,5 @@
 package org.sterl.spring.persistent_tasks.trigger.component;
 
-import java.time.OffsetDateTime;
 import java.util.Optional;
 
 import org.springframework.lang.Nullable;
@@ -9,8 +8,6 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.sterl.spring.persistent_tasks.api.task.RunningTriggerContextHolder;
 import org.sterl.spring.persistent_tasks.task.TaskService;
-import org.sterl.spring.persistent_tasks.task.exception.CancelTaskException;
-import org.sterl.spring.persistent_tasks.task.exception.FailTaskNoRetryException;
 import org.sterl.spring.persistent_tasks.trigger.model.RunTaskWithStateCommand;
 import org.sterl.spring.persistent_tasks.trigger.model.TriggerEntity;
 
@@ -23,6 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 public class RunTriggerComponent {
 
     private final TaskService taskService;
+    private final FailTriggerComponent failTrigger;
     private final EditTriggerComponent editTrigger;
     private final StateSerializer serializer = new StateSerializer();
 
@@ -43,7 +41,7 @@ public class RunTriggerComponent {
             RunningTriggerContextHolder.setContext(runTaskWithState.runningTrigger());
             return runTaskWithState.execute(editTrigger);
         } catch (Exception e) {
-            return failTaskAndState(runTaskWithState, e);
+            return failTrigger.execute(runTaskWithState, e);
         } finally {
             RunningTriggerContextHolder.clearContext();
         }
@@ -57,37 +55,8 @@ public class RunTriggerComponent {
             final var state = serializer.deserialize(trigger.getData().getState());
             return new RunTaskWithStateCommand(task, trx, state, trigger);
         } catch (Exception e) {
-            failTaskAndState(new RunTaskWithStateCommand(null, Optional.empty(), null, trigger), e);
+            failTrigger.execute(trigger, e);
             return null;
         }
-    }
-    
-    private Optional<TriggerEntity> failTaskAndState(RunTaskWithStateCommand runTaskWithStateCommand, Exception e) {
-
-        var trigger = runTaskWithStateCommand.trigger();
-        var task = runTaskWithStateCommand.task();
-        Optional<TriggerEntity> result;
-
-        if (e instanceof CancelTaskException) {
-            log.info("Cancel of a running trigger={} requested", trigger.getKey());
-            result = editTrigger.cancelTask(trigger.getKey(), e);
-        } else if (e instanceof FailTaskNoRetryException) {
-            log.warn("Fail no retry of a running trigger={} requested", trigger.getKey(), e);
-            result = editTrigger.failTrigger(trigger.getKey(), runTaskWithStateCommand.state(), e, null);
-        } else if (task == null 
-                || !task.retryStrategy().shouldRetry(trigger.getData().getExecutionCount(), e)) {
-
-            log.error("Failed trigger={}, no further retries!", trigger.getKey(), e);
-            result = editTrigger.failTrigger(trigger.getKey(), runTaskWithStateCommand.state(), e, null);
-        } else {
-            final OffsetDateTime retryAt = task.retryStrategy().retryAt(trigger.getData().getExecutionCount(), e);
-            if (retryAt == null) {
-                log.error("Failed trigger={}, no further retries!", trigger.getKey(), e);
-            } else {
-                log.warn("Failed trigger={} with retryAt={}", trigger.getKey(), retryAt, e);
-            }
-            result = editTrigger.failTrigger(trigger.getKey(), runTaskWithStateCommand.state(), e, retryAt);
-        }
-        return result;
     }
 }